[Standalone] Some fixes wrt. the shared_ptr transition
Note that comparisons on boost::shared_ptr ar basedd on the control_block,
whereas comparisons on std::shared_ptr are based on the pointee.
This way, multiple boost::shared_ptr which has been initialized with
a pointer of value zero are still comparing unequal to each other
and e.g. a default constructed boost::shared_ptr.